feat: добавить логирование при удалении пользователей и фильмов

+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -14,6 +14,7 @@ import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.mapper.EventMapper;
 import ru.yandex.practicum.filmorate.mapper.FilmMapper;
 import ru.yandex.practicum.filmorate.mapper.UserMapper;
+import ru.yandex.practicum.filmorate.model.Director;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.model.events.Event;
@@ -257,6 +258,15 @@ public class UserService {
     }
 
     public void removeUserById(int userId) {
+        logger.debug("Запрос на удаление пользователя с id = {}", userId);
+
+        Optional<User> maybeUser = userRepository.getById(userId);
+        if (maybeUser.isEmpty()) {
+            logger.warn("Пользователь с id = {} не найден", userId);
+            throw new NotFoundException("Пользователь с id = " + userId + " не найден");
+        }
+
         userRepository.removeUserById(userId);
+        logger.debug("Удалён пользователь с id = {}", userId);
     }
 }
